ResultSet expects schema only when job is complete and supports requesting tokens when required (#3)
* ResultSet schema expection bug fix

The schema field of `QueryResponse` is only present when the query job has completed. Therefore, if not complete we should not expect a schema in `ResultSet`.

Co-Authored-By: Rob Fraser <rob.fraser@bigpayme.com>

* Fetch access tokens from the authenticator when required

This commit updates the APIs to request tokens from the authenticator when required instead of storing a single token that may expire. `yup_oauth2` will cache tokens and automatically refresh or re-request tokens when they have expired.

Co-Authored-By: Rob Fraser <rob.fraser@bigpayme.com>

Co-authored-by: Rob Fraser <rob.fraser@bigpayme.com>